+++ springboot/src/main/java/com/mzc/backend/lms/domains/message/conversation/entity/Conversation.java
@@ -1,6 +1,6 @@
 package com.mzc.backend.lms.domains.message.conversation.entity;
 
-import com.mzc.backend.lms.domains.user.user.entity.User;
+import com.mzc.backend.lms.domains.user.adapter.out.persistence.entity.User;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
