+++ springboot/src/main/java/com/mzc/backend/lms/domains/message/message/dto/MessageResponseDto.java
@@ -1,7 +1,7 @@
 package com.mzc.backend.lms.domains.message.message.dto;
 
 import com.mzc.backend.lms.domains.message.message.entity.Message;
-import com.mzc.backend.lms.domains.user.auth.encryption.annotation.Encrypted;
+import com.mzc.backend.lms.domains.user.encryption.annotation.Encrypted;
 import lombok.Builder;
 import lombok.Getter;
