+++ springboot/src/main/java/com/mzc/backend/lms/domains/message/message/service/MessageService.java
@@ -11,9 +11,9 @@ import com.mzc.backend.lms.domains.message.message.entity.Message;
 import com.mzc.backend.lms.domains.message.message.repository.MessageRepository;
 import com.mzc.backend.lms.domains.message.sse.dto.MessageNotificationDto;
 import com.mzc.backend.lms.domains.message.sse.service.SseService;
-import com.mzc.backend.lms.domains.user.profile.entity.UserProfile;
-import com.mzc.backend.lms.domains.user.user.entity.User;
-import com.mzc.backend.lms.domains.user.user.repository.UserRepository;
+import com.mzc.backend.lms.domains.user.adapter.out.persistence.entity.UserProfile;
+import com.mzc.backend.lms.domains.user.adapter.out.persistence.entity.User;
+import com.mzc.backend.lms.domains.user.adapter.out.persistence.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.PageRequest;
